Fix different platform handling
In the case a page is not for the `common` platform, we would mistakenly
show the error message that a page is for a different platform while the
plaform would be correct. We should only show the error when the page is
not for the `common` and current platform.

The cache was also updated before checking for a alternative platform.
This is now changed to update after checking if a page is not found
because it may be unnecessary and would occur to often which caused
annoying delays.